+++ packages/did-ui-react/src/components/config-provider/types.ts
@@ -1,5 +1,5 @@
 import { IConfig } from '@portkey/types';
-import { ISocialLoginConfig, TotalAccountType } from '../../types';
+import { ISocialLoginConfig, TCustomNetworkType, TSupportAccountType } from '../../types';
 import { BaseReCaptcha } from '../types';
 
 export interface ConfigProviderProps {
@@ -7,11 +7,9 @@ export interface ConfigProviderProps {
 }
 
 export interface ILoginConfig {
-  loginMethodsOrder?: TotalAccountType[];
+  loginMethodsOrder?: TSupportAccountType[];
   recommendIndexes?: number[];
 }
-
-export type TCustomNetworkType = 'Offline' | 'onLine';
 
 export interface GlobalConfigProps extends IConfig {
   socialLogin?: ISocialLoginConfig;