+++ packages/did-ui-react/src/components/config-provider/utils.ts
@@ -1,5 +1,5 @@
 import ConfigProvider from './';
-import { dealURLLastChar } from '../../utils';
+import { dealURLLastChar, did } from '../../utils';
 
 export const getServiceUrl = () => {
   if (!ConfigProvider.config.serviceUrl) throw Error('Please config serviceUrl');
@@ -12,4 +12,10 @@ export const getSocketUrl = () => {
 
 export const getCustomNetworkType = () => {
   return ConfigProvider.config.customNetworkType || 'onLine';
+};
+
+export const getStorageInstance = () => {
+  const storageMethod = ConfigProvider.config.storageMethod || did.config.storageMethod;
+  if (!storageMethod) throw Error('Please config storageMethod');
+  return storageMethod;
 };