+++ packages/next-example/src/app/assets/page.tsx
@@ -5,7 +5,6 @@ import React from 'react';
 
 ConfigProvider.setGlobalConfig({
   // storageMethod: myStore,
-  serviceUrl: 'https://localtest-applesign2.portkey.finance',
 
   requestDefaults: {
     timeout: 30000,
@@ -18,7 +17,7 @@ ConfigProvider.setGlobalConfig({
 
 export default function Assets() {
   return (
-    <PortkeyAssetProvider pin="111111" originChainId="tDVW">
+    <PortkeyAssetProvider pin="111111" originChainId="AELF">
       <Asset
         faucet={{
           faucetContractAddress: '2UM9eusxdRyCztbmMZadGXzwgwKfFdk8pF4ckw58D769ehaPSR',