+++ packages/socket/src/index.ts
@@ -1,5 +1,6 @@
 export * from './didSignalr';
 export * from './signalr';
 export * from './sellSignalr';
+export * from './openloginSignalr';
 
 export * from './types';